// WADNR.Web/src/app/shared/generated/model/forester-work-unit-grid-row.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ForesterWorkUnitGridRow { 
    ForesterWorkUnitID?: number;
    ForesterRoleID?: number;
    ForesterRoleDisplayName?: string | null;
    ForesterWorkUnitName?: string | null;
    RegionName?: string | null;
    PersonID?: number | null;
    AssignedPersonName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ForesterWorkUnitGridRowForm { 
    ForesterWorkUnitID?: FormControl<number>;
    ForesterRoleID?: FormControl<number>;
    ForesterRoleDisplayName?: FormControl<string>;
    ForesterWorkUnitName?: FormControl<string>;
    RegionName?: FormControl<string>;
    PersonID?: FormControl<number>;
    AssignedPersonName?: FormControl<string>;
}

export class ForesterWorkUnitGridRowFormControls { 
    public static ForesterWorkUnitID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ForesterRoleID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ForesterRoleDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ForesterWorkUnitName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static RegionName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static PersonID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static AssignedPersonName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}